[GREEN] Create CouponService interface and modify existing implementation and test accordingly
